Playing over localhost works with one player.
With two players there is a nullpointerexception.
Foot animation is not yet hooked up yet
Bullets need to be put in some kind of behavior for them to work.
Items list in the map is broken because it cannot be created. The items might not exist when the list is created